an algorithm for data compression
A version of the Ziv-Lempel data compression algorithm is the basis for must popular compression programs such as winzip, zip or gzip. Implemented a hash table class using open addressing with linear probing. Hash table is general and works for any arbitrary object class that overloads the operator == and/or !=.